Struct aws_sdk_robomaker::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for robomaker.

This client allows ergonomic access to a robomaker-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_delete_worlds(&self) -> BatchDeleteWorlds<C, M, R>

pub fn batch_describe_simulation_job(
&self
) -> BatchDescribeSimulationJob<C, M, R>

pub fn cancel_deployment_job(&self) -> CancelDeploymentJob<C, M, R>

pub fn cancel_simulation_job(&self) -> CancelSimulationJob<C, M, R>

pub fn cancel_simulation_job_batch(&self) -> CancelSimulationJobBatch<C, M, R>

pub fn cancel_world_export_job(&self) -> CancelWorldExportJob<C, M, R>

pub fn cancel_world_generation_job(&self) -> CancelWorldGenerationJob<C, M, R>

pub fn create_deployment_job(&self) -> CreateDeploymentJob<C, M, R>

pub fn create_fleet(&self) -> CreateFleet<C, M, R>

pub fn create_robot(&self) -> CreateRobot<C, M, R>

pub fn create_robot_application(&self) -> CreateRobotApplication<C, M, R>

pub fn create_robot_application_version(
&self
) -> CreateRobotApplicationVersion<C, M, R>

pub fn create_simulation_application(
&self
) -> CreateSimulationApplication<C, M, R>

pub fn create_simulation_application_version(
&self
) -> CreateSimulationApplicationVersion<C, M, R>

pub fn create_simulation_job(&self) -> CreateSimulationJob<C, M, R>

pub fn create_world_export_job(&self) -> CreateWorldExportJob<C, M, R>

pub fn create_world_generation_job(&self) -> CreateWorldGenerationJob<C, M, R>

pub fn create_world_template(&self) -> CreateWorldTemplate<C, M, R>

pub fn delete_fleet(&self) -> DeleteFleet<C, M, R>

pub fn delete_robot(&self) -> DeleteRobot<C, M, R>

pub fn delete_robot_application(&self) -> DeleteRobotApplication<C, M, R>

pub fn delete_simulation_application(
&self
) -> DeleteSimulationApplication<C, M, R>

pub fn delete_world_template(&self) -> DeleteWorldTemplate<C, M, R>

pub fn deregister_robot(&self) -> DeregisterRobot<C, M, R>

pub fn describe_deployment_job(&self) -> DescribeDeploymentJob<C, M, R>

pub fn describe_fleet(&self) -> DescribeFleet<C, M, R>

pub fn describe_robot(&self) -> DescribeRobot<C, M, R>

pub fn describe_robot_application(&self) -> DescribeRobotApplication<C, M, R>

pub fn describe_simulation_application(
&self
) -> DescribeSimulationApplication<C, M, R>

pub fn describe_simulation_job(&self) -> DescribeSimulationJob<C, M, R>

pub fn describe_simulation_job_batch(
&self
) -> DescribeSimulationJobBatch<C, M, R>

pub fn describe_world(&self) -> DescribeWorld<C, M, R>

pub fn describe_world_export_job(&self) -> DescribeWorldExportJob<C, M, R>

pub fn describe_world_generation_job(
&self
) -> DescribeWorldGenerationJob<C, M, R>

pub fn describe_world_template(&self) -> DescribeWorldTemplate<C, M, R>

pub fn get_world_template_body(&self) -> GetWorldTemplateBody<C, M, R>

pub fn list_deployment_jobs(&self) -> ListDeploymentJobs<C, M, R>

pub fn list_fleets(&self) -> ListFleets<C, M, R>

pub fn list_robot_applications(&self) -> ListRobotApplications<C, M, R>

pub fn list_robots(&self) -> ListRobots<C, M, R>

pub fn list_simulation_applications(
&self
) -> ListSimulationApplications<C, M, R>

pub fn list_simulation_job_batches(&self) -> ListSimulationJobBatches<C, M, R>

pub fn list_simulation_jobs(&self) -> ListSimulationJobs<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_world_export_jobs(&self) -> ListWorldExportJobs<C, M, R>

pub fn list_world_generation_jobs(&self) -> ListWorldGenerationJobs<C, M, R>

pub fn list_worlds(&self) -> ListWorlds<C, M, R>

pub fn list_world_templates(&self) -> ListWorldTemplates<C, M, R>

pub fn register_robot(&self) -> RegisterRobot<C, M, R>

pub fn restart_simulation_job(&self) -> RestartSimulationJob<C, M, R>

pub fn start_simulation_job_batch(&self) -> StartSimulationJobBatch<C, M, R>

pub fn sync_deployment_job(&self) -> SyncDeploymentJob<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_robot_application(&self) -> UpdateRobotApplication<C, M, R>

pub fn update_simulation_application(
&self
) -> UpdateSimulationApplication<C, M, R>

pub fn update_world_template(&self) -> UpdateWorldTemplate<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self