Module aws_sdk_robomaker::error

Modules

concurrent_deployment_exception

See ConcurrentDeploymentException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

internal_server_exception

See InternalServerException

invalid_parameter_exception

See InvalidParameterException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

throttling_exception

See ThrottlingException

Structs

BatchDeleteWorldsError

BatchDescribeSimulationJobError

CancelDeploymentJobError

CancelSimulationJobBatchError

CancelSimulationJobError

CancelWorldExportJobError

CancelWorldGenerationJobError

ConcurrentDeploymentException

The failure percentage threshold percentage was met.

CreateDeploymentJobError

CreateFleetError

CreateRobotApplicationError

CreateRobotApplicationVersionError

CreateRobotError

CreateSimulationApplicationError

CreateSimulationApplicationVersionError

CreateSimulationJobError

CreateWorldExportJobError

CreateWorldGenerationJobError

CreateWorldTemplateError

DeleteFleetError

DeleteRobotApplicationError

DeleteRobotError

DeleteSimulationApplicationError

DeleteWorldTemplateError

DeregisterRobotError

DescribeDeploymentJobError

DescribeFleetError

DescribeRobotApplicationError

DescribeRobotError

DescribeSimulationApplicationError

DescribeSimulationJobBatchError

DescribeSimulationJobError

DescribeWorldError

DescribeWorldExportJobError

DescribeWorldGenerationJobError

DescribeWorldTemplateError

GetWorldTemplateBodyError

IdempotentParameterMismatchException

The request uses the same client token as a previous, but non-identical request. Do not

InternalServerException

AWS RoboMaker experienced a service issue. Try your call again.

InvalidParameterException

A parameter specified in a request is not valid, is unsupported, or cannot be used. The

LimitExceededException

The requested resource exceeds the maximum number allowed, or the number of concurrent

ListDeploymentJobsError

ListFleetsError

ListRobotApplicationsError

ListRobotsError

ListSimulationApplicationsError

ListSimulationJobBatchesError

ListSimulationJobsError

ListTagsForResourceError

ListWorldExportJobsError

ListWorldGenerationJobsError

ListWorldTemplatesError

ListWorldsError

RegisterRobotError

ResourceAlreadyExistsException

The specified resource already exists.

ResourceNotFoundException

The specified resource does not exist.

RestartSimulationJobError

ServiceUnavailableException

The request has failed due to a temporary failure of the server.

StartSimulationJobBatchError

SyncDeploymentJobError

TagResourceError

ThrottlingException

AWS RoboMaker is temporarily unable to process the request. Try your call again.

UntagResourceError

UpdateRobotApplicationError

UpdateSimulationApplicationError

UpdateWorldTemplateError

Enums

BatchDeleteWorldsErrorKind

BatchDescribeSimulationJobErrorKind

CancelDeploymentJobErrorKind

CancelSimulationJobBatchErrorKind

CancelSimulationJobErrorKind

CancelWorldExportJobErrorKind

CancelWorldGenerationJobErrorKind

CreateDeploymentJobErrorKind

CreateFleetErrorKind

CreateRobotApplicationErrorKind

CreateRobotApplicationVersionErrorKind

CreateRobotErrorKind

CreateSimulationApplicationErrorKind

CreateSimulationApplicationVersionErrorKind

CreateSimulationJobErrorKind

CreateWorldExportJobErrorKind

CreateWorldGenerationJobErrorKind

CreateWorldTemplateErrorKind

DeleteFleetErrorKind

DeleteRobotApplicationErrorKind

DeleteRobotErrorKind

DeleteSimulationApplicationErrorKind

DeleteWorldTemplateErrorKind

DeregisterRobotErrorKind

DescribeDeploymentJobErrorKind

DescribeFleetErrorKind

DescribeRobotApplicationErrorKind

DescribeRobotErrorKind

DescribeSimulationApplicationErrorKind

DescribeSimulationJobBatchErrorKind

DescribeSimulationJobErrorKind

DescribeWorldErrorKind

DescribeWorldExportJobErrorKind

DescribeWorldGenerationJobErrorKind

DescribeWorldTemplateErrorKind

GetWorldTemplateBodyErrorKind

ListDeploymentJobsErrorKind

ListFleetsErrorKind

ListRobotApplicationsErrorKind

ListRobotsErrorKind

ListSimulationApplicationsErrorKind

ListSimulationJobBatchesErrorKind

ListSimulationJobsErrorKind

ListTagsForResourceErrorKind

ListWorldExportJobsErrorKind

ListWorldGenerationJobsErrorKind

ListWorldTemplatesErrorKind

ListWorldsErrorKind

RegisterRobotErrorKind

RestartSimulationJobErrorKind

StartSimulationJobBatchErrorKind

SyncDeploymentJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateRobotApplicationErrorKind

UpdateSimulationApplicationErrorKind

UpdateWorldTemplateErrorKind