Struct aws_sdk_robomaker::model::SimulationJob

#[non_exhaustive]
pub struct SimulationJob {

    pub arn: Option<String>,
    pub name: Option<String>,
    pub status: Option<SimulationJobStatus>,
    pub last_started_at: Option<Instant>,
    pub last_updated_at: Option<Instant>,
    pub failure_behavior: Option<FailureBehavior>,
    pub failure_code: Option<SimulationJobErrorCode>,
    pub failure_reason: Option<String>,
    pub client_request_token: Option<String>,
    pub output_location: Option<OutputLocation>,
    pub logging_config: Option<LoggingConfig>,
    pub max_job_duration_in_seconds: i64,
    pub simulation_time_millis: i64,
    pub iam_role: Option<String>,
    pub robot_applications: Option<Vec<RobotApplicationConfig>>,
    pub simulation_applications: Option<Vec<SimulationApplicationConfig>>,
    pub data_sources: Option<Vec<DataSource>>,
    pub tags: Option<HashMap<String, String>>,
    pub vpc_config: Option<VpcConfigResponse>,
    pub network_interface: Option<NetworkInterface>,
    pub compute: Option<ComputeResponse>,

}

Information about a simulation job.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

arn: Option<String>

The Amazon Resource Name (ARN) of the simulation job.

name: Option<String>

The name of the simulation job.

status: Option<SimulationJobStatus>

Status of the simulation job.

last_started_at: Option<Instant>

The time, in milliseconds since the epoch, when the simulation job was last started.

last_updated_at: Option<Instant>

The time, in milliseconds since the epoch, when the simulation job was last updated.

failure_behavior: Option<FailureBehavior>

The failure behavior the simulation job.

Continue

Leaves the host running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

failure_code: Option<SimulationJobErrorCode>

The failure code of the simulation job if it failed.

failure_reason: Option<String>

The reason why the simulation job failed.

client_request_token: Option<String>

A unique identifier for this SimulationJob request.

output_location: Option<OutputLocation>

Location for output files generated by the simulation job.

logging_config: Option<LoggingConfig>

The logging configuration.

max_job_duration_in_seconds: i64

The maximum simulation job duration in seconds. The value must be 8 days (691,200 seconds) or less.

simulation_time_millis: i64

The simulation job execution duration in milliseconds.

iam_role: Option<String>

The IAM role that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

robot_applications: Option<Vec<RobotApplicationConfig>>

A list of robot applications.

simulation_applications: Option<Vec<SimulationApplicationConfig>>

A list of simulation applications.

data_sources: Option<Vec<DataSource>>

The data sources for the simulation job.

tags: Option<HashMap<String, String>>

A map that contains tag keys and tag values that are attached to the simulation job.

vpc_config: Option<VpcConfigResponse>

VPC configuration information.

network_interface: Option<NetworkInterface>

Information about a network interface.

compute: Option<ComputeResponse>

Compute information for the simulation job

Implementations

impl SimulationJob

pub fn builder() -> Builder

Creates a new builder-style object to manufacture SimulationJob