Module aws_sdk_robomaker::operation

Structs

BatchDeleteWorlds

Deletes one or more worlds in a batch operation.

BatchDescribeSimulationJob

Describes one or more simulation jobs.

CancelDeploymentJob

Cancels the specified deployment job.

CancelSimulationJob

Cancels the specified simulation job.

CancelSimulationJobBatch

Cancels a simulation job batch. When you cancel a simulation job batch, you are also

CancelWorldExportJob

Cancels the specified export job.

CancelWorldGenerationJob

Cancels the specified world generator job.

CreateDeploymentJob

Deploys a specific version of a robot application to robots in a fleet.

CreateFleet

Creates a fleet, a logical group of robots running the same robot application.

CreateRobot

Creates a robot.

CreateRobotApplication

Creates a robot application.

CreateRobotApplicationVersion

Creates a version of a robot application.

CreateSimulationApplication

Creates a simulation application.

CreateSimulationApplicationVersion

Creates a simulation application with a specific revision id.

CreateSimulationJob

Creates a simulation job.

CreateWorldExportJob

Creates a world export job.

CreateWorldGenerationJob

Creates worlds using the specified template.

CreateWorldTemplate

Creates a world template.

DeleteFleet

Deletes a fleet.

DeleteRobot

Deletes a robot.

DeleteRobotApplication

Deletes a robot application.

DeleteSimulationApplication

Deletes a simulation application.

DeleteWorldTemplate

Deletes a world template.

DeregisterRobot

Deregisters a robot.

DescribeDeploymentJob

Describes a deployment job.

DescribeFleet

Describes a fleet.

DescribeRobot

Describes a robot.

DescribeRobotApplication

Describes a robot application.

DescribeSimulationApplication

Describes a simulation application.

DescribeSimulationJob

Describes a simulation job.

DescribeSimulationJobBatch

Describes a simulation job batch.

DescribeWorld

Describes a world.

DescribeWorldExportJob

Describes a world export job.

DescribeWorldGenerationJob

Describes a world generation job.

DescribeWorldTemplate

Describes a world template.

GetWorldTemplateBody

Gets the world template body.

ListDeploymentJobs

Returns a list of deployment jobs for a fleet. You can optionally provide filters to

ListFleets

Returns a list of fleets. You can optionally provide filters to retrieve specific

ListRobotApplications

Returns a list of robot application. You can optionally provide filters to retrieve

ListRobots

Returns a list of robots. You can optionally provide filters to retrieve specific

ListSimulationApplications

Returns a list of simulation applications. You can optionally provide filters to

ListSimulationJobBatches

Returns a list simulation job batches. You can optionally provide filters to retrieve

ListSimulationJobs

Returns a list of simulation jobs. You can optionally provide filters to retrieve

ListTagsForResource

Lists all tags on a AWS RoboMaker resource.

ListWorldExportJobs

Lists world export jobs.

ListWorldGenerationJobs

Lists world generator jobs.

ListWorldTemplates

Lists world templates.

ListWorlds

Lists worlds.

RegisterRobot

Registers a robot with a fleet.

RestartSimulationJob

Restarts a running simulation job.

StartSimulationJobBatch

Starts a new simulation job batch. The batch is defined using one or more

SyncDeploymentJob

Syncrhonizes robots in a fleet to the latest deployment. This is helpful if robots were

TagResource

Adds or edits tags for a AWS RoboMaker resource.

UntagResource

Removes the specified tags from the specified AWS RoboMaker resource.

UpdateRobotApplication

Updates a robot application.

UpdateSimulationApplication

Updates a simulation application.

UpdateWorldTemplate

Updates a world template.